requirements.txt
langgraph==1.0.1
langchain==1.0.2
langchain-core==1.0.0
langchain-ollama==1.0.0
ollama==0.6.0

lark-oapi==1.4.23

aiohttp==3.13.1
requests==2.32.5

pyyaml==6.0.3

# 需要C编译器支持
#    - Windows: 安装 Visual Studio Build Tools
#    - macOS: xcode-select --install
#    - Linux: sudo apt-get install build-essential
tree-sitter==0.21.3
tree-sitter-languages==1.10.2


# ripgrep==15.1.0
# 需要单独安装，不在pip中：
#   Windows:  choco install ripgrep
#   macOS:    brew install ripgrep
#   Linux:    sudo apt-get install ripgrep
